finalize tests
