--- tests/test_input_params.py
@@ -1,2 +1,0 @@
-
-
